Added Encoders, Added Gyro (NO TESTING)
Added the encoders to the drivetrain and derived a forumla (linear so
thus proportional) for the distance traveled in inches from the pulse
value.
d(pulse) = 3pi * (p/(3*256))

Also included metric convesions :)

New Gyro subsystem using a PID controller stolen from a previous working
NavX gyro test on my personal GitHub. (The implementation here is yet to
be tuned for the robot)